// src/app/model/bulk-transfer-consumer.model.ts
export interface BulkTransferConsumer {
    custId: string | null;
    locationCode: string | null;
    customerNumber: string | null;
    customerName: string | null;
    customerAddr: string | null;
    locationName: string | null;
    nidNumber: string | null;
    mobileNumber: string | null;
    powerUtility: string | null;
    maxPower: string | null;
    fatherName: string | null;
    address: string | null;
    tariff: string | null;
    BusinessType: string | null;
    prvAcNo: string | null;
    areaCode: string | null;
    division: string | null;
    district: string | null;
    thana: string | null;
    meterType: string | null;
    meterNum: string | null;
    meterCondition: string | null;
    meterOwner: string | null;
    sanctionLoad: string | null;
    billNumber: string | null;
    locationDeptCode: string | null;
    tariffId: number | null;
    businessType: string | null;
    tariffDesc: string | null;
    meterTypeDesc: string | null;
    meterTypeCode: string | null;
    meterConditionCode: string | null;
    meterConditionDesc: string | null;
    businessTypeCode: string | null;
    billReasonCode: string | null;
    multypliBy: string | null;
    dcTypeCode: string | null;
}